separates components
