+++ components/index.js
@@ -15,6 +15,8 @@ import ButtonGroup from './ButtonGroup';
 import CarouselExample from './CarouselExample';
 import SegmentedControlExample from './SegmentedControlExample';
 import FooterTagsList from './FooterTagsList';
+import HorizontalMenuDefault from './HorizontalMenuDefault'
+import CommentsBoxWithForm from './CommentsBoxWithForm'
 
 
 export {
@@ -35,4 +37,6 @@ export {
   CarouselExample,
   SegmentedControlExample,
   FooterTagsList,
+  HorizontalMenuDefault,
+  CommentsBoxWithForm,
 }
